+++ molo/core/content_import/utils.py
@@ -14,3 +14,10 @@ def conj(a, b):
 
 def omit_nones(d):
     return dict((k, v) for k, v in d.iteritems() if v is not None)
+
+
+def schedule(task, delay=True, *a, **kw):
+    if delay:
+        return task.delay(*a, **kw)
+    else:
+        return task(*a, **kw)