+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/settings/production.py
@@ -44,16 +44,16 @@ COMPRESS_OFFLINE = True
 # Setup for CAS
 ENABLE_SSO = True
 
-MIDDLEWARE_CLASSES += (
+MIDDLEWARE_CLASSES += [
     'molo.core.middleware.MoloCASMiddleware',
     'molo.core.middleware.Custom403Middleware',
-)
+]
 
 
-AUTHENTICATION_BACKENDS = (
+AUTHENTICATION_BACKENDS = [
     'django.contrib.auth.backends.ModelBackend',
     'molo.core.backends.MoloCASBackend',
-)
+]
 
 CAS_SERVER_URL = ''
 CAS_ADMIN_PREFIX = '/admin/'